separate persistence perf testkit + extracted BatchingSqlJournal
