+++ src/contrib/persistence/Akka.Persistence.Sqlite.Tests.Performance/Akka.Persistence.Sqlite.Tests.Performance.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{453EFD22-7C53-4887-9DBF-FCFC9172E909}</ProjectGuid>
+    <ProjectGuid>{6ACC5CB4-AB86-4587-8C5B-C5713E25295B}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Akka.Persistence.Sqlite</RootNamespace>
-    <AssemblyName>Akka.Persistence.Sqlite</AssemblyName>
+    <RootNamespace>Akka.Persistence.Sqlite.Tests.Performance</RootNamespace>
+    <AssemblyName>Akka.Persistence.Sqlite.Tests.Performance</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <NuGetPackageImportStamp>
@@ -30,15 +30,13 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <DocumentationFile>bin\Release\Akka.Persistence.Sqlite.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Collections.Immutable, Version=1.1.36.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-      <HintPath>..\..\..\packages\System.Collections.Immutable.1.1.36\lib\portable-net45+win8+wp8+wpa81\System.Collections.Immutable.dll</HintPath>
+    <Reference Include="NBench, Version=0.3.1.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\..\..\packages\NBench.0.3.1\lib\net45\NBench.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="System.Configuration" />
+    <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Data.SQLite, Version=1.0.103.0, Culture=neutral, PublicKeyToken=db937bc2d44ff139, processorArchitecture=MSIL">
       <HintPath>..\..\..\packages\System.Data.SQLite.Core.1.0.103\lib\net45\System.Data.SQLite.dll</HintPath>
@@ -52,25 +50,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConnectionContext.cs" />
-    <Compile Include="Extension.cs" />
-    <Compile Include="Journal\BatchingSqliteJournal.cs" />
-    <Compile Include="Journal\BatchingSqliteJournalSettings.cs" />
-    <Compile Include="Journal\BatchingSqlJournal.cs" />
-    <Compile Include="Journal\MultiValueDictionaryExtensions.cs" />
-    <Compile Include="Journal\SqliteQueryExecutor.cs" />
-    <Compile Include="Journal\SqliteJournal.cs" />
+    <Compile Include="BatchingJournal\BatchingSqliteManyActorsRecoverySpec.cs" />
+    <Compile Include="BatchingJournal\BatchingSqliteManyActorsWriteSpec.cs" />
+    <Compile Include="BatchingJournal\BatchingSqliteRawJournalWriteSpec.cs" />
+    <Compile Include="BatchingJournal\BatchingSqliteSingleActorWriteSpec.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Snapshot\SqliteSnapshotStore.cs" />
+    <Compile Include="SqliteManyActorsRecoverySpec.cs" />
+    <Compile Include="SqliteManyActorsWriteSpec.cs" />
+    <Compile Include="SqliteRawJournalWriteSpec.cs" />
+    <Compile Include="SqliteSingleActorWriteSpec.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="Akka.Persistence.Sqlite.nuspec" />
-    <None Include="packages.config" />
-    <EmbeddedResource Include="sqlite.conf">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
+    <ProjectReference Include="..\..\..\core\Akka.Persistence.Performance.TestKit\Akka.Persistence.Performance.TestKit.csproj">
+      <Project>{5e7d8489-07ef-4fa1-b880-16a2e32bde59}</Project>
+      <Name>Akka.Persistence.Performance.TestKit</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\..\core\Akka.Persistence\Akka.Persistence.csproj">
       <Project>{fca84dea-c118-424b-9eb8-34375dfef18a}</Project>
       <Name>Akka.Persistence</Name>
@@ -83,8 +77,23 @@
       <Project>{3b9e6211-9488-4db5-b714-24248693b38f}</Project>
       <Name>Akka.Persistence.Sql.Common</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Akka.Persistence.Sqlite\Akka.Persistence.Sqlite.csproj">
+      <Project>{453efd22-7c53-4887-9dbf-fcfc9172e909}</Project>
+      <Name>Akka.Persistence.Sqlite</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup />
+  <ItemGroup>
+    <None Include="packages.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <Import Project="..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets" Condition="Exists('..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets')" />
+  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
+    <PropertyGroup>
+      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
+    </PropertyGroup>
+    <Error Condition="!Exists('..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets'))" />
+  </Target>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
   <Target Name="BeforeBuild">
@@ -92,12 +101,4 @@
   <Target Name="AfterBuild">
   </Target>
   -->
-  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
-    <PropertyGroup>
-      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
-    </PropertyGroup>
-    <Error Condition="!Exists('..\..\..\packages\System.Data.SQLite.Core.1.0.98.1\build\net45\System.Data.SQLite.Core.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\..\..\packages\System.Data.SQLite.Core.1.0.98.1\build\net45\System.Data.SQLite.Core.targets'))" />
-    <Error Condition="!Exists('..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets'))" />
-  </Target>
-  <Import Project="..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets" Condition="Exists('..\..\..\packages\System.Data.SQLite.Core.1.0.103\build\net45\System.Data.SQLite.Core.targets')" />
 </Project>
