+++ src/contrib/persistence/Akka.Persistence.Sqlite.Tests.Performance/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NBench" version="0.3.1" targetFramework="net45" />
+  <package id="System.Data.SQLite.Core" version="1.0.103" targetFramework="net45" />
+</packages>
