+++ src/core/Akka.Persistence.Performance.TestKit/Properties/AssemblyInfo.cs
