+++ src/core/Akka.Persistence.Performance.TestKit/packages.config
